Calculate the precision, recall, and f-scores of overlaps between two GRanges objects indicating enhancer-gene relations.

Given two GRanges objects, each of which has the meta column named "gene", this function calculates the overlaps between them. Based on the calculated overlaps, this function computes precision, recall, and f-score. This function aims to assess the enhancer-gene relations in eRegulons or eGRNs. We may use the enhancer-target pair databases, e.g., EnhancerAtlas or scEnhancer of the same tissues or cell lines.

intersect_enhancer_gene_relations(x, y)

Arguments

x

The first GRanges object saving enhancer-gene relations with gene symbols saved in "gene" meta column

y

The second GRanges object saving enhancer-gene relations with gene symbols saved in "gene" meta column

Value

Return a data.frame indicating overlapped GRanges objects, containing the following columns:

x.peak: The enhancer in the first GRanges object for each pair of overlapped GRanges objects.

y.peak: The enhancer in the second GRanges object for each pair of overlapped GRanges objects.

gene: The gene for each pair of overlapped GRanges objects.

References

Li, Y., Ma, A., Wang, Y., Wang, C., Chen, S., Fu, H., Liu, B. and Ma, Q., 2022. Enhancer-driven gene regulatory networks inference from single-cell RNA-seq and ATAC-seq data. bioRxiv, pp.2022-12.

Gao, T., & Qian, J. (2020). EnhancerAtlas 2.0: an updated resource with enhancer annotation in 586 tissue/cell types across nine species. Nucleic acids research, 48(D1), D58-D64.

Gao, T., Zheng, Z., Pan, Y., Zhu, C., Wei, F., Yuan, J., ... & Qian, J. (2022). scEnhancer: a single-cell enhancer resource with annotation across hundreds of tissue/cell types in three species. Nucleic acids research, 50(D1), D371-D379.