Get the dotplot-heatmap for cell-type-specific enhancer regulons (eRegulons)

Get the dotplot-heatmap for cell-type-specific enhancer regulons (eRegulons)

plot_dotplot_heatmap(
  obj = NULL,
  cts.en.regs = NULL,
  n.regs = 5,
  celltype.col = "seurat_clusters",
  dot = "AC",
  peak.assay = "ATAC",
  ht.colours = c((grDevices::colorRampPalette(c("#99D5FF", "#E8F9FD")))(3), "#fffaf7",
    (grDevices::colorRampPalette(c("#ffe4e1", "#ee7261")))(3)),
  ht.breaks = NULL,
  font.size = 15,
  text.font = "Arial",
  angle = 30,
  celltypes = NULL,
  vjust = 1,
  hjust = 1,
  celltype.colours = NULL
)

Arguments

obj

A Seurat object, NULL by default

cts.en.regs

The list of cell-type-specific enhancer regulons (eRegulons)

n.regs

The number of top-ranked eRegulons for plotting, 5 by default

celltype.col

The metadata column indicating cell types, "seurat_clusters" by default

dot

What dot sizes represent, "AC" by default

peak.assay

The assay denoting chromatin accessibility, "ATAC" by default

ht.colours

Color list used to define the gradients of the heatmap, c( grDevices::colorRampPalette(c("#99D5FF", "#E8F9FD"))(3), "#fffaf7", grDevices::colorRampPalette(c("#ffe4e1", "#ee7261"))(3) ) by default

ht.breaks

The breaks marked on the legend of heatmap colors

font.size

The font size of legends

text.font

The text font of legends

angle

The angle of labels of the x-axis in the heatmap, 30 by default

celltypes

The list of cell types arranged in an order defined by users

vjust

Adjustment of positions in the vertical direction, 1 by default

hjust

Adjustment of positions in the horizontal direction, 1 by default

celltype.colours

The list of colors of cell types in the row annotations